Added support to leave comments in InfoDialog.js. But for now the comments aren't stored in ContextBroker, since the http://orion.lab.fi-ware.org server wont authorize with a current token. POST 401 (Unauthorized) error message is printed on browser console when request is being send to the server.
For now the comment GUI code was tested using custom DebugPoiStorage class that simulates the read and write of the context broker. DebugPoiStorage class can be found at the begining of the InfoDialog.js file. If you want to swap back to Context Broker, you can uncomment "getPoiComments" (line 201) and "createPoiComment" (line 222) lines in InfoDialog.js.